requirements.txt
# Backend requirements
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.5.0
python-multipart==0.0.6
sentence-transformers==2.3.1
transformers==4.36.0
huggingface_hub==0.19.4
torch==2.1.1
numpy==1.24.4
qdrant-client==1.7.0
google-generativeai==0.3.2
pypdf==3.17.1
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.25.2
